revert: rename to readDoc
so github can actually track these changes properly 🙃
