+++ src/lib/readDoc.ts
