+++ src/lib/syncDocsPath.legacy.ts
@@ -1,4 +1,4 @@
-import type { PageMetadata } from './readPage.js';
+import type { PageMetadata } from './readDoc.js';
 import type ChangelogsCommand from '../commands/changelogs.js';
 
 import fs from 'node:fs/promises';
@@ -9,8 +9,8 @@ import toposort from 'toposort';
 
 import { APIv1Error } from './apiError.js';
 import readdirRecursive from './readdirRecursive.js';
+import readPage from './readDoc.js';
 import { cleanAPIv1Headers, handleAPIv1Res, readmeAPIv1Fetch } from './readmeAPIFetch.js';
-import readPage from './readPage.js';
 
 /** API path within ReadMe to update (e.g. `docs`, `changelogs`, etc.) */
 type PageType = 'changelogs' | 'custompages' | 'docs';
